API authentication

+++ src/modules/message/MessageApi.ts
@@ -1,6 +1,8 @@
 import { NextFunction, Request, RequestHandler, Response, Router } from 'express';
 import { Container } from 'typedi';
 
+import * as Auth from '../../middlewares/auth.middleware';
+
 import Random from '../../utils/random';
 
 import MessageService from './MessageService';
@@ -9,10 +11,10 @@ const create: RequestHandler = async (req: Request, res: Response, next: NextFun
     console.debug('Calling insert message: %o', req.body);
     try{
         const service = Container.get<MessageService>(MessageService);
-        const { content, senderId, recipientId, replyTo, roomId } = req.body;
+        const { content, recipientId, replyTo, roomId } = req.body;
         const id = Random.id();
 
-        await service.create(id, content, senderId, recipientId, replyTo, roomId);
+    await service.create(id, content, req.user.id, recipientId, replyTo, roomId);
 
         return res.status(200). json({ id });
     } catch(e){
@@ -26,6 +28,11 @@ const deleteMessage: RequestHandler = async (req: Request, res: Response, next: 
     try{
         const service = Container.get<MessageService>(MessageService);
         const { id } = req.body;
+        const message = await service.get(id);
+
+        if(message.senderId !== req.user.id){
+            return res.status(401).json({ message: 'Not enough privileges' });
+        }
 
         const deleted = await service.delete(id);
 
@@ -37,6 +44,6 @@ const deleteMessage: RequestHandler = async (req: Request, res: Response, next: 
 };
 
 export default (app: Router): void => {
-    app.post('/message/', create);
-    app.delete('/message/', deleteMessage);
+  app.post('/message/', Auth.authorize([]), create);
+  app.delete('/message/', Auth.authorize([]), deleteMessage);
 };
